Disabling and re-enabling of network activity in case we encounter timeouts

+++ code/app/src/main/java/com/tanfra/shopmob/smob/data/repo/SmobGroupRepository.kt
@@ -42,6 +42,8 @@ class SmobGroupRepository(
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : SmobGroupDataSource, KoinComponent {
 
+    // fetch worker class form service locator
+    private val wManager: SmobAppWork by inject()
 
     // --- impl. of public, app facing data interface 'SmobGroupDataSource': CRUD, local DB data ---
     // --- impl. of public, app facing data interface 'SmobGroupDataSource': CRUD, local DB data ---
@@ -114,13 +116,15 @@ class SmobGroupRepository(
 
             // then push to backend DB
             // ... PUT or POST? --> try a GET first to find out if item already exists in backend DB
-            val testRead = getSmobGroupViaApi(dbGroup.id)
-            if (testRead.data?.id != dbGroup.id) {
-                // item not found in backend --> use POST to create it
-                saveSmobGroupViaApi(dbGroup)
-            } else {
-                // item already exists in backend DB --> use PUT to update it
-                smobGroupApi.updateSmobGroupById(dbGroup.id, dbGroup.asNetworkModel())
+            if(wManager.netActive) {
+                val testRead = getSmobGroupViaApi(dbGroup.id)
+                if (testRead.data?.id != dbGroup.id) {
+                    // item not found in backend --> use POST to create it
+                    saveSmobGroupViaApi(dbGroup)
+                } else {
+                    // item already exists in backend DB --> use PUT to update it
+                    smobGroupApi.updateSmobGroupById(dbGroup.id, dbGroup.asNetworkModel())
+                }
             }
 
         }  // idlingResource (testing)
@@ -154,10 +158,9 @@ class SmobGroupRepository(
 
                 // then push to backend DB
                 // ... use 'update', as group may already exist (equivalent of REPLACE w/h local DB)
-                //
-                // ... could do a read back first, if we're anxious...
-                //smobGroupDao.getSmobGroupById(dbGroup.id)?.let { smobGroupApi.updateSmobGroup(it.id, it.asNetworkModel()) }
-                smobGroupApi.updateSmobGroupById(dbGroup.id, dbGroup.asNetworkModel())
+                if(wManager.netActive) {
+                    smobGroupApi.updateSmobGroupById(dbGroup.id, dbGroup.asNetworkModel())
+                }
 
             }
         }
@@ -182,7 +185,9 @@ class SmobGroupRepository(
             // support espresso testing (w/h coroutines)
             wrapEspressoIdlingResource {
                 smobGroupDao.deleteSmobGroupById(id)
-                smobGroupApi.deleteSmobGroupById(id)
+                if(wManager.netActive) {
+                    smobGroupApi.deleteSmobGroupById(id)
+                }
             }
         }
     }
@@ -199,11 +204,13 @@ class SmobGroupRepository(
                 smobGroupDao.deleteAllSmobGroups()
 
                 // then delete all groups from backend DB
-                getSmobGroupsViaApi().let {
-                    if (it.status == Status.SUCCESS) {
-                        it.data?.map { smobGroupApi.deleteSmobGroupById(it.id) }
-                    } else {
-                        Timber.w("Unable to get SmobGroup IDs from backend DB (via API) - not deleting anything.")
+                if(wManager.netActive) {
+                    getSmobGroupsViaApi().let {
+                        if (it.status == Status.SUCCESS) {
+                            it.data?.map { smobGroupApi.deleteSmobGroupById(it.id) }
+                        } else {
+                            Timber.w("Unable to get SmobGroup IDs from backend DB (via API) - not deleting anything.")
+                        }
                     }
                 }
             }
