+++ code/app/src/main/java/com/tanfra/shopmob/smob/data/repo/SmobProductRepository.kt
@@ -14,6 +14,7 @@ import com.tanfra.shopmob.smob.data.net.nto2dto.asRepoModel
 import com.tanfra.shopmob.smob.data.repo.utils.Resource
 import com.tanfra.shopmob.smob.data.repo.utils.Status
 import com.tanfra.shopmob.smob.data.repo.utils.asResource
+import com.tanfra.shopmob.smob.work.SmobAppWork
 import com.tanfra.shopmob.utils.wrapEspressoIdlingResource
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.Flow
@@ -39,6 +40,8 @@ class SmobProductRepository(
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : SmobProductDataSource, KoinComponent {
 
+    // fetch worker class form service locator
+    private val wManager: SmobAppWork by inject()
 
     // --- impl. of public, app facing data interface 'SmobProductDataSource': CRUD, local DB data ---
     // --- impl. of public, app facing data interface 'SmobProductDataSource': CRUD, local DB data ---
@@ -138,13 +141,15 @@ class SmobProductRepository(
 
             // then push to backend DB
             // ... PUT or POST? --> try a GET first to find out if item already exists in backend DB
-            val testRead = getSmobProductViaApi(dbProduct.id)
-            if (testRead.data?.id != dbProduct.id) {
-                // item not found in backend --> use POST to create it
-                saveSmobProductViaApi(dbProduct)
-            } else {
-                // item already exists in backend DB --> use PUT to update it
-                smobProductApi.updateSmobProductById(dbProduct.id, dbProduct.asNetworkModel())
+            if(wManager.netActive) {
+                val testRead = getSmobProductViaApi(dbProduct.id)
+                if (testRead.data?.id != dbProduct.id) {
+                    // item not found in backend --> use POST to create it
+                    saveSmobProductViaApi(dbProduct)
+                } else {
+                    // item already exists in backend DB --> use PUT to update it
+                    smobProductApi.updateSmobProductById(dbProduct.id, dbProduct.asNetworkModel())
+                }
             }
 
         }  // idlingResource (testing)
@@ -178,10 +183,9 @@ class SmobProductRepository(
 
                 // then push to backend DB
                 // ... use 'update', as product may already exist (equivalent of REPLACE w/h local DB)
-                //
-                // ... could do a read back first, if we're anxious...
-                //smobProductDao.getSmobProductById(dbProduct.id)?.let { smobProductApi.updateSmobProduct(it.id, it.asNetworkModel()) }
-                smobProductApi.updateSmobProductById(dbProduct.id, dbProduct.asNetworkModel())
+                if(wManager.netActive) {
+                    smobProductApi.updateSmobProductById(dbProduct.id, dbProduct.asNetworkModel())
+                }
 
             }
         }
@@ -206,7 +210,9 @@ class SmobProductRepository(
             // support espresso testing (w/h coroutines)
             wrapEspressoIdlingResource {
                 smobProductDao.deleteSmobProductById(id)
-                smobProductApi.deleteSmobProductById(id)
+                if(wManager.netActive) {
+                    smobProductApi.deleteSmobProductById(id)
+                }
             }
         }
     }
@@ -223,11 +229,13 @@ class SmobProductRepository(
                 smobProductDao.deleteAllSmobProducts()
 
                 // then delete all products from backend DB
-                getSmobProductsViaApi().let {
-                    if (it.status == Status.SUCCESS) {
-                        it.data?.map { smobProductApi.deleteSmobProductById(it.id) }
-                    } else {
-                        Timber.w("Unable to get SmobProduct IDs from backend DB (via API) - not deleting anything.")
+                if(wManager.netActive) {
+                    getSmobProductsViaApi().let {
+                        if (it.status == Status.SUCCESS) {
+                            it.data?.map { smobProductApi.deleteSmobProductById(it.id) }
+                        } else {
+                            Timber.w("Unable to get SmobProduct IDs from backend DB (via API) - not deleting anything.")
+                        }
                     }
                 }
             }
