+++ code/app/src/main/java/com/tanfra/shopmob/smob/data/repo/SmobUserRepository.kt
@@ -14,6 +14,7 @@ import com.tanfra.shopmob.smob.data.net.nto2dto.asRepoModel
 import com.tanfra.shopmob.smob.data.repo.utils.Resource
 import com.tanfra.shopmob.smob.data.repo.utils.Status
 import com.tanfra.shopmob.smob.data.repo.utils.asResource
+import com.tanfra.shopmob.smob.work.SmobAppWork
 import com.tanfra.shopmob.utils.wrapEspressoIdlingResource
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.Flow
@@ -39,6 +40,8 @@ class SmobUserRepository(
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : SmobUserDataSource, KoinComponent {
 
+    // fetch worker class form service locator
+    private val wManager: SmobAppWork by inject()
 
     // --- impl. of public, app facing data interface 'SmobUserDataSource': CRUD, local DB data ---
     // --- impl. of public, app facing data interface 'SmobUserDataSource': CRUD, local DB data ---
@@ -110,11 +113,17 @@ class SmobUserRepository(
             smobUserDao.saveSmobUser(dbUser)
 
             // then push to backend DB
-            // ... use 'update', as shop may already exist (equivalent of REPLACE w/h local DB)
-            //
-            // ... could do a read back first, if we're anxious...
-            //smobUserDao.getSmobUserById(dbUser.id)?.let { smobUserApi.updateSmobUser(it.id, it.asNetworkModel()) }
-            smobUserApi.updateSmobUserById(dbUser.id, dbUser.asNetworkModel())
+            // ... PUT or POST? --> try a GET first to find out if item already exists in backend DB
+            if(wManager.netActive) {
+                val testRead = getSmobUserViaApi(dbUser.id)
+                if (testRead.data?.id != dbUser.id) {
+                    // item not found in backend --> use POST to create it
+                    saveSmobUserViaApi(dbUser)
+                } else {
+                    // item already exists in backend DB --> use PUT to update it
+                    smobUserApi.updateSmobUserById(dbUser.id, dbUser.asNetworkModel())
+                }
+            }
 
         }  // idlingResource (testing)
 
@@ -147,13 +156,15 @@ class SmobUserRepository(
 
                 // then push to backend DB
                 // ... PUT or POST? --> try a GET first to find out if item already exists in backend DB
-                val testRead = getSmobUserViaApi(dbUser.id)
-                if (testRead.data?.id != dbUser.id) {
-                    // item not found in backend --> use POST to create it
-                    saveSmobUserViaApi(dbUser)
-                } else {
-                    // item already exists in backend DB --> use PUT to update it
-                    smobUserApi.updateSmobUserById(dbUser.id, dbUser.asNetworkModel())
+                if(wManager.netActive) {
+                    val testRead = getSmobUserViaApi(dbUser.id)
+                    if (testRead.data?.id != dbUser.id) {
+                        // item not found in backend --> use POST to create it
+                        saveSmobUserViaApi(dbUser)
+                    } else {
+                        // item already exists in backend DB --> use PUT to update it
+                        smobUserApi.updateSmobUserById(dbUser.id, dbUser.asNetworkModel())
+                    }
                 }
 
             }
@@ -179,7 +190,9 @@ class SmobUserRepository(
             // support espresso testing (w/h coroutines)
             wrapEspressoIdlingResource {
                 smobUserDao.deleteSmobUserById(id)
-                smobUserApi.deleteSmobUserById(id)
+                if(wManager.netActive) {
+                    smobUserApi.deleteSmobUserById(id)
+                }
             }
         }
     }
@@ -196,11 +209,13 @@ class SmobUserRepository(
                 smobUserDao.deleteAllSmobUsers()
 
                 // then delete all shops from backend DB
-                getSmobUsersViaApi().let {
-                    if (it.status == Status.SUCCESS) {
-                        it.data?.map { smobUserApi.deleteSmobUserById(it.id) }
-                    } else {
-                        Timber.w("Unable to get SmobUser IDs from backend DB (via API) - not deleting anything.")
+                if(wManager.netActive) {
+                    getSmobUsersViaApi().let {
+                        if (it.status == Status.SUCCESS) {
+                            it.data?.map { smobUserApi.deleteSmobUserById(it.id) }
+                        } else {
+                            Timber.w("Unable to get SmobUser IDs from backend DB (via API) - not deleting anything.")
+                        }
                     }
                 }
             }
